python/flask/rest_docstrings/docs/requirements.txt
sphinx-rtd-theme==0.1.6
sphinxcontrib-httpdomain==1.3.0
sphinxcontrib-napoleon==0.2.8